requirements.txt
python==3.7.3
pytorch==1.7.1
numpy==1.19.4
scikit-learn==0.23.2
scipy==1.5.4